FIX-#2080: engine dispatching moved to a separate folder
